--- pr_test_20250703_021835.md
@@ -1,16 +1,0 @@
-# PR 테스트 파일
-
-**생성 시간**: 2025-07-03 02:18:35
-**파일명**: pr_test_20250703_021835.md
-
-## 목적
-GitHub PR 생성 테스트를 위한 파일입니다.
-
-## 테스트 내용
-- Git 커밋 테스트
-- Git 푸시 테스트  
-- PR URL 생성 테스트
-- 브라우저 열기 테스트
-
-## 참고사항
-이 파일은 테스트 후 삭제해도 됩니다.
